+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 # 0.13.2
-
+- Fix for "Fix including nested directories".
 
 # 0.13.1
 - Code coverage for internal simulator.